restructurazon del proyecto parte 1
