--- src/components/forms/LoginForm.tsx
@@ -1,5 +1,0 @@
-import React from 'react'
-
-export default function LoginForm() {
-  return <div>LoginForm</div>
-}
