+++ src/components/headers/HeaderPage.tsx
@@ -1,10 +1,10 @@
 import * as React from 'react'
 import { Box, Toolbar, Typography, AppBar } from '@mui/material'
-import CodeIcon from '@mui/icons-material/Code'
 import MenuItem from '@mui/material/MenuItem'
 import AccountCircle from '@mui/icons-material/AccountCircle'
 import IconButton from '@mui/material/IconButton'
 import Menu from '@mui/material/Menu'
+import { Logo } from '../../utils/const/icons'
 
 export default function HeaderPage() {
   const [anchorEl, setAnchorEl] = React.useState<null | HTMLElement>(null)
@@ -18,11 +18,10 @@ export default function HeaderPage() {
     <Box sx={{ flexGrow: 1, width: 1, flexWrap: 'wrap' }}>
       <AppBar position='static' sx={{ backgroundColor: 'common.white' }}>
         <Toolbar>
-          <CodeIcon sx={{ mr: 1 }} />
+          <Logo sx={{ mr: 1 }} />
           <Typography variant='h6' component='div' sx={{ flexGrow: 1 }}>
             Logo and Brand
           </Typography>
-
           <IconButton
             size='large'
             aria-label='account of current user'
