--- src/components/headers/Headerlogin/HeaderLogin.tsx
@@ -1,68 +1,0 @@
-import * as React from 'react'
-import AppBar from '@mui/material/AppBar'
-import Box from '@mui/material/Box'
-import Toolbar from '@mui/material/Toolbar'
-import Typography from '@mui/material/Typography'
-import IconButton from '@mui/material/IconButton'
-import CodeIcon from '@mui/icons-material/Code'
-import AccountCircle from '@mui/icons-material/AccountCircle'
-import MenuItem from '@mui/material/MenuItem'
-import Menu from '@mui/material/Menu'
-
-export default function HeaderLogin() {
-  const [anchorEl, setAnchorEl] = React.useState(false)
-  const anchorRef = React.useRef<HTMLButtonElement>(null)
-
-  const handleToggle = () => {
-    setAnchorEl((prevOpen) => !prevOpen)
-  }
-
-  const handleClose = (event: Event | React.SyntheticEvent) => {
-    if (anchorRef.current && anchorRef.current.contains(event.target as HTMLElement)) {
-      return
-    }
-
-    setAnchorEl(false)
-  }
-  return (
-    <Box sx={{ flexGrow: 1, width: 1, height: '10%', flexWrap: 'wrap' }}>
-      <AppBar position='static' sx={{ backgroundColor: 'common.white' }}>
-        <Toolbar>
-          <CodeIcon sx={{ mr: 1 }} />
-          <Typography variant='h6' component='div' sx={{ flexGrow: 1 }}>
-            Logo and Brand
-          </Typography>
-          <div>
-            <IconButton
-              ref={anchorRef}
-              size='large'
-              aria-label='account of current user'
-              aria-controls='menu-appbar'
-              aria-haspopup='true'
-              onClick={handleToggle}
-              color='inherit'>
-              <AccountCircle />
-            </IconButton>
-            <Menu
-              id='menu-appbar'
-              anchorOrigin={{
-                vertical: 'top',
-                horizontal: 'right',
-              }}
-              keepMounted
-              transformOrigin={{
-                vertical: 'top',
-                horizontal: 'right',
-              }}
-              open={Boolean(anchorEl)}
-              onClose={handleClose}>
-              <MenuItem onClick={handleClose}>Profile</MenuItem>
-              <MenuItem onClick={handleClose}>My account</MenuItem>
-              <MenuItem onClick={handleClose}>Logout</MenuItem>
-            </Menu>
-          </div>
-        </Toolbar>
-      </AppBar>
-    </Box>
-  )
-}
