+++ src/pages/Dashboard.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Container } from '@mui/material'
-import AsideBar from '../components/drawers/AsideBar'
+import AsideBar from '../components/Drawer/AsideBar'
 import HomeCourse from '../components/HomeLogin/HomeCourse'
 export default function Log() {
   return (
@@ -9,8 +9,6 @@ export default function Log() {
       sx={{
         height: '100vh',
         padding: '0 !important',
-        backgroundColor: 'green',
-        display: 'flex',
       }}>
       <AsideBar />
       <HomeCourse />
